requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.0
bleach==6.1.0
blinker==1.7.0
bottle==0.12.25
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coverage==7.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.7
Django==4.2.6
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-compressor==4.4
django-haystack==3.2.1
django-ipware==5.0.1
django-mdeditor==0.1.20
django-uuslug==2.0.0
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
filelock==3.12.4
Flask==3.0.2
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.3
jsonpickle==3.0.2
lxml==5.1.0
Markdown==3.5
MarkupSafe==2.1.5
multidict==6.0.4
mysqlclient==2.2.0
oauthlib==3.2.2
openai==0.28.1
passlib==1.7.4
Pillow==10.0.1
platformdirs==3.10.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-logstash==0.4.8
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.4
sqlparse==0.4.3
text-unidecode==1.3
tqdm==4.66.1
tzdata==2022.7
ua-parser==0.18.0
urllib3==1.26.17
user-agents==2.2.0
virtualenv==20.24.5
webencodings==0.5.1
Werkzeug==3.0.1
WeRoBot==1.13.1
Whoosh==2.7.4
xmltodict==0.13.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
